pyproject.toml
[project]
name = "kerberos-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "ThePhilAz",email = "rami.azouz@philico.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "requests-kerberos (>=0.15.0,<0.16.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "gssapi (>=1.10.0,<2.0.0)",
    "pykerberos (>=1.2.4,<2.0.0)",
    "ruff (>=0.13.0,<0.14.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
]

[project.scripts]
uq_web = "test_unique_web_search:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
